+++ internal/provider/helper.go
@@ -58,20 +58,12 @@ func validateAutoscallingLimit(val interface{}, name string) (warns []string, er
 	switch val.(type) {
 	case float64:
 		switch v := val.(float64); v {
-		case 0.25,
-			0.5,
-			1.,
-			2.,
-			3.,
-			4.,
-			5.,
-			6.,
-			7.:
+		case 0.25, 0.5, 1., 2., 3., 4., 5., 6., 7., 8., 9., 10.:
 			return
 		}
 	case int:
 		switch v := val.(int); v {
-		case 1, 2, 3, 4, 5, 6, 7:
+		case 1, 2, 3, 4, 5, 6, 7, 8, 9, 10:
 			return
 		}
 	}