+++ internal/provider/helper_test.go
@@ -12,7 +12,7 @@ func Test_validateAutoscallingLimit(t *testing.T) {
 
 	t.Run(
 		"happy path: int input", func(t *testing.T) {
-			input := []int{1, 2, 3, 4, 5, 6, 7}
+			input := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 			for _, in := range input {
 				_, errs := validateAutoscallingLimit(in, "")
 				if errs != nil {
@@ -24,7 +24,7 @@ func Test_validateAutoscallingLimit(t *testing.T) {
 
 	t.Run(
 		"happy path: float64 input", func(t *testing.T) {
-			input := []float64{0.25, 0.5, 1, 2, 3, 4, 5, 6, 7}
+			input := []float64{0.25, 0.5, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 			for _, in := range input {
 				_, errs := validateAutoscallingLimit(in, "")
 				if errs != nil {
@@ -36,7 +36,7 @@ func Test_validateAutoscallingLimit(t *testing.T) {
 
 	t.Run(
 		"unhappy path", func(t *testing.T) {
-			input := []interface{}{"foo", 0, 0.1, 1.5, 10}
+			input := []interface{}{"foo", 0, 0.1, 1.5, 11, 12, 20}
 			for _, in := range input {
 				_, errs := validateAutoscallingLimit(in, "")
 				if errs == nil {